+++ src/services/storage/gcp.ts
@@ -1,5 +1,6 @@
 import { Storage } from '@google-cloud/storage';
 import { IStorageService } from '.';
+import { generatePublicUri } from '../../config';
 
 /**
  * Google Cloud Storage service.
@@ -31,6 +32,9 @@ export class GCPStorageService implements IStorageService {
 
         await file.save(body, options);
 
+        const publicUri = generatePublicUri(key);
+        if (publicUri) return { uri: publicUri };
+
         return { uri: `https://${bucket}.storage.googleapis.com/${key}` };
     }
 
@@ -43,12 +47,7 @@ export class GCPStorageService implements IStorageService {
     async objectExists(bucket: string, key: string): Promise<boolean> {
         const bucketInstance = this.storage.bucket(bucket);
         const file = bucketInstance.file(key);
-
-        try {
-            const [exists] = await file.exists();
-            return exists;
-        } catch (error) {
-            return false;
-        }
+        const [exists] = await file.exists();
+        return exists;
     }
 }